ARM: Flush the caches for non panicking CPUs in case of a kernel panic
In case of a kernel panic, only the panicking CPU does an entire
cache flush. This means that certain dirty cache lines in the L1
caches of the other CPUs may never get flushed. This gives us
improper RAM dumps. Add cache flushing for all the online CPUs.
The outer domain is not flushed since it is already being done by
the panicking CPU.

Change-Id: I64fe79587de6afc612d228f78c2f71e031d91f89
Signed-off-by: Syed Rameez Mustafa <rameezmustafa@codeaurora.org>