cpu_input_boost: Add power-saving algorithm to selectively boost CPUs
Instead of boosting all CPUs, CPU input boost now checks two factors when determining the amount of CPUs to boost: the number of CPUs that are online and the current frequency of online CPUs. Each of these factors can only queue up to 2 CPUs for boosting, and a maximum of 3 out of 4 CPUs can boosted at once (or 2 CPUs on dual-core systems). A new sysfs parameter, up_threshold, has been added which can tweak the aggressiveness of the first half of the power algorithm (the current frequency of each online CPU).

Online CPUs are given priority when boosting.

Change-Id: I36117a2d2c46087282d36ca0fcb7258c43d9f47a
Signed-off-by: Sultanxda <sultanxda@gmail.com>